Replaced precision.h to avoid segfault

+++ src/precision.h
@@ -1,10 +1,10 @@
-/*
- *  precision.h
- *  samurai
- *
- *  Copyright 2011 Michael M. Bell
- *  All rights reserved.
- *
+/*                                                                                                                         
+ *  precision.h                                                                                                            
+ *  samurai                                                                                                                
+ *                                                                                                                         
+ *  Copyright 2011 Michael M. Bell                                                                                         
+ *  All rights reserved.                                                                                                   
+ *                                                                                                                         
  */
 
 #ifndef PRECISION_H
@@ -13,6 +13,7 @@
 #include <cstdint>
 
 typedef double real;
-typedef unsigned long int integer;
+//typedef unsigned long int integer;                                                                                       
+typedef unsigned int integer;
 
 #endif
